Move package from staff to about
